Add a little extra JS to fix panel flow
This began as a CSS-only tab situation, but it has had JS for so long
to manage URL hashes that it might as well use a little more to fix
the layout.  By managing hidden/shown state in JS instead of CSS, we
can move the panels away from being siblings of their tabs, which
fixes layout.